refactor(wpf-migration phase 0a): extract dependency-free static helpers to Core
Moves 9 pure (System.IO/string/Linq, no DotNetZip/crypto) static helpers out of
Main.cs into Savedrake.Core, keeping the WinForms app building via thin forwarders
and the harness green. No behavior change.

- SaveScan: IsRealSaveEntry, FindDd2SaveFoldersUnder (+ Dd2AppId/DirLastWriteUtc),
  FindLatestPreRestoreCheckpoint, BackupLocationWarning
- BackupNaming: MakeUniquePath
- Pinning: PinTag, IsPinnedBackup, PinnedPath, UnpinnedPath
- RetentionPolicy: SelectAutobackupsToThin
- Main.cs keeps private forwarders; harness repoints these 7 lookups to Core types.

Verified: Debug + Release build clean; harness 196/196 (the moved methods now run
against Core via the repointed lookups, so the behavior is proven intact).
